Removing user game feature restored. Seperated games and usergames class functions. Dashboard now refreshes datatable after removing games.
