+++ src/main/java/fish/payara/hello/jsf/UserGamesBean.java
@@ -1,26 +1,29 @@
 package fish.payara.hello.jsf;
 
 import fish.payara.hello.entities.Games;
-import fish.payara.hello.entities.UserGames;
+import fish.payara.hello.entities.UserAccount;
 import fish.payara.hello.service.UserGamesService;
 import jakarta.annotation.PostConstruct;
+import jakarta.faces.context.FacesContext;
 import jakarta.faces.view.ViewScoped;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
+import org.primefaces.PrimeFaces;
+
+import java.io.IOException;
 import java.io.Serializable;
-import java.util.*;
+import java.util.List;
 
 @Named(value = "userGamesBean")
 @ViewScoped
 public class UserGamesBean implements Serializable {
 
-    private int gameId;
-    private int userId;
-
     private List<Games> games;
 
+    private int userId;
+
     @Inject
-    private UserGamesService service;
+    private UserGamesService userGamesService;
 
     @Inject
     private LoginBean loginBean;
@@ -28,30 +31,34 @@ public class UserGamesBean implements Serializable {
     @Inject
     private UserAccountBean userAccountBean;
 
-    public UserGamesBean() {
-        //for JPA
-    }
-
     @PostConstruct
     public void init() {
         userId = userAccountBean.getLoggedInUserId(loginBean.getUsername());
-        games = service.listAllGamesInDashboard(userId);
+        games = userGamesService.listAllGamesInDashboard(userId);
     }
 
-    public int getGameId() {
-        return gameId;
+    //need to redirect to login page if user is not logged in
+    public void saveGameToDashboard(Games game) throws IOException {
+        FacesContext facesContext = FacesContext.getCurrentInstance();
+        UserAccount user = (UserAccount) facesContext.getExternalContext().getSessionMap().get("user");
+        if (user == null) {
+            facesContext.getExternalContext().redirect("login.xhtml");
+            return;
+        }
+        userGamesService.saveGameToDashboard(user, game);
     }
 
-    public void setGameId(int gameId) {
-        this.gameId = gameId;
-    }
+    public void removeGameFromDashboard(Games game) {
+        UserAccount user = userAccountBean.getUser(userId);
+        if (user != null && game != null) {
+            userGamesService.removeGameFromDashboard(user, game);
 
-    public int getUserId() {
-        return userId;
-    }
+            // Refresh the list after removal
+            games = userGamesService.listAllGamesInDashboard(user.getId());
 
-    public void setUserId(int userId) {
-        this.userId = userId;
+            // Optionally update UI components
+            PrimeFaces.current().ajax().update("gameTable");
+        }
     }
 
     public List<Games> getGames() {
