+++ src/main/java/fish/payara/hello/service/GameService.java
@@ -25,36 +25,4 @@ public class GameService {
     @PersistenceContext
     EntityManager em;
 
-    //TODO: stop user from saving the same game to their dashboard, handle this
-    /*
-    stateless = public = transations
-     */
-    public void saveGameToDashboard(UserAccount user, Games game) {
-        if (game == null) {
-            return;
-        }
-        UserGames checkForExisting = null;
-        try {
-            checkForExisting = em.createNamedQuery("UserGames.findByUserIdAndGameId", UserGames.class)
-                    .setParameter("userId", user.getId())
-                    .setParameter("gameId", game.getId())
-                    .getSingleResult();
-        } catch (NoResultException e) {
-            Logger.getLogger(GameService.class.getName()).log(Level.INFO, "No existing game found for user, proceeding to save new game", e);
-        }
-        if (checkForExisting == null) {
-            UserGames userGames = new UserGames(user, game.getId());
-            em.persist(userGames);
-        }
-    }
-//        // add primeface notification pop up if saved already within else block, mb change ui for save game button to show it already saved with primefaces
-//        Logger.getLogger(GameService.class.getName()).log(Level.INFO, "Game already saved to the dashboard");
-
-    public void removeGameFromDashboard(UserAccount user, Games game) {
-//        UserGames userGames = em.createNamedQuery("UserGames.findByUserIdAndGameId", UserGames.class)
-//                .setParameter("userId", user.getId())
-//                .setParameter("gameId", game.getId())
-//                .getSingleResult();
-//        em.remove(userGames);
-    }
 }
