+++ src/main/java/fish/payara/hello/service/UserGamesService.java
@@ -8,12 +8,15 @@ import jakarta.ejb.Stateless;
 import jakarta.ejb.LocalBean;
 import jakarta.inject.Inject;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.NoResultException;
 import jakarta.persistence.PersistenceContext;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  *
@@ -40,5 +43,39 @@ public class UserGamesService {
             games.addAll(igdbService.getGameByID(userGame.getGame()));
         }
         return games;
+    }
+
+    public void saveGameToDashboard(UserAccount user, Games game) {
+        if (game == null) {
+            return;
+        }
+        UserGames checkForExisting = null;
+        try {
+            checkForExisting = em.createNamedQuery("UserGames.findByUserIdAndGameId", UserGames.class)
+                    .setParameter("userId", user.getId())
+                    .setParameter("gameId", game.getId())
+                    .getSingleResult();
+        } catch (NoResultException e) {
+            Logger.getLogger(GameService.class.getName()).log(Level.INFO, "No existing game found for user, proceeding to save new game", e);
+        }
+        if (checkForExisting == null) {
+            UserGames userGames = new UserGames(user, game.getId());
+            em.persist(userGames);
+        }
+    }
+//        // add primeface notification pop up if saved already within else block, mb change ui for save game button to show it already saved with primefaces
+//        Logger.getLogger(GameService.class.getName()).log(Level.INFO, "Game already saved to the dashboard");
+
+    public void removeGameFromDashboard(UserAccount user, Games game) {
+        try {
+            UserGames userGames = em.createNamedQuery("UserGames.findByUserIdAndGameId", UserGames.class)
+                    .setParameter("userId", user.getId())
+                    .setParameter("gameId", game.getId())
+                    .getSingleResult();
+            em.remove(userGames);
+        } catch (NoResultException e) {
+            System.out.println("No game found for user: " + user.getId() + " and game: " + game.getId());
+
+        }
     }
 }
