+++ src/main/webapp/index.xhtml
@@ -48,36 +48,11 @@
                 </f:facet>
                 <h:outputText value="Critic Rating: #{game.aggregated_rating}"/>
                 <f:facet name="footer">
-                    <p:commandButton icon="pi pi-check" value="Save" action="#{gameBean.saveGameToDashboard(game)}"/>
+                    <p:commandButton icon="pi pi-check" value="Save" action="#{userGamesBean.saveGameToDashboard(game)}"/>
                 </f:facet>
             </p:card>
         </ui:repeat>
         </div>
     </h:form>
-
-
-    <!--        DISPLAY ALL GAMES in datatable format TODO: reuse this for search functionality -->
-<!--    <h:form>-->
-<!--        <p:dataTable value="#{gameBean.games}" var="game" styleClass="game-table">-->
-<!--            <p:column headerText="Title" sortBy="#{game.name}">-->
-<!--                <h:outputText value="#{game.name}"/>-->
-<!--            </p:column>-->
-<!--            <p:column headerText="Genres">-->
-<!--                <h:outputText value="#{game.genres}"/>-->
-<!--            </p:column>-->
-<!--            <p:column headerText="First Release Date" sortBy="#{game.releaseDate}">-->
-<!--                <h:outputText value="#{game.releaseDate}"/>-->
-<!--            </p:column>-->
-<!--            <p:column headerText="Rating" sortBy="#{game.rating}">-->
-<!--                <h:outputText value="#{game.rating.intValue()}"/>-->
-<!--            </p:column>-->
-<!--            <p:column>-->
-<!--                <f:facet name="header">-->
-<!--                    <h:outputText value="Save"/>-->
-<!--                </f:facet>-->
-<!--                <p:commandButton value="+" action="#{gameBean.saveGameToDashboard(game)}" styleClass="game-save-button"/>-->
-<!--            </p:column>-->
-<!--        </p:dataTable>-->
-<!--    </h:form>-->
 </h:body>
 </html>
